requirements.txt
Django>=5.0,<6.0
django-allauth>=64.0.0
djangorestframework>=3.15.0
django-filter>=24.2
Pillow>=10.0.0
python-dotenv>=1.0.1
# Optional / later integrations (commented until needed)
# stripe>=10.0.0
# boto3>=1.34.0
# celery>=5.4.0
# redis>=5.0.0